Explicitly cast quality to int

+++ rulesets/mason/world/tasks/Sift.py
@@ -83,10 +83,10 @@ class Sift(server.Task):
             moisture = 1
         self_loc.coordinates = self.pos
 
-        quality = 10 * self.get_quality(self_loc.coordinates, target, moisture)
-        for i in range(quality/2, quality):
+        quality = int(10 * self.get_quality(self_loc.coordinates, target, moisture))
+        for i in range(int(quality/2), quality):
             res = res + Operation("create", Entity(name = "scrawny earthworm", parents = ["annelid"], location = self_loc), to=self.character)
-        for i in range((10-quality)/2, quality):
+        for i in range(int((10-quality)/2), quality):
             res = res + Operation("create", Entity(name = "juicy earthworm", parents = ["annelid"], location = self_loc), to=self.character)
 
         #res.append(self.next_tick(1.75))
